adding dirty hack

+++ conn/src/rc.cpp
@@ -496,21 +496,21 @@ void *ReliableConnection ::getLocalSetup() {
   printf("===== LOCAL ADDRESS : %p\n", reinterpret_cast<void*>(addr));
   printf("===== LOCAL KEY : %p\n\n", reinterpret_cast<void*>(lkey));
 
-  memcpy(static_cast<uint16_t*>(privateData) + 4, reinterpret_cast<void*>(&addr), 8);   // Address
-  memcpy(static_cast<uint16_t*>(privateData) + 20, reinterpret_cast<void*>(&lkey), 4);  // Key
+  memcpy(static_cast<int*>(privateData) + 4, reinterpret_cast<void*>(&addr), 8);   // Address
+  memcpy(static_cast<int*>(privateData) + 20, reinterpret_cast<void*>(&lkey), 4);  // Key
 
   return privateData;
 }
 
 void ReliableConnection ::setRemoteSetup(const void *network_data) {
   // 4 Bytes of offset to get the address
-  memcpy(&rconn.rci.buf_addr, static_cast<uint16_t*>(network_data) + 4, 8);
+  memcpy(&rconn.rci.buf_addr, static_cast<int*>(network_data) + 4, 8);
 
   rconn.rci.buf_size = mr.size; //en supposant que tous les buffers ont la même taille partout 
   //on copie la taille de notre local buffer à nous 
 
   // 20 Bytes of offset to get KEY
-  memcpy(&rconn.rci.rkey, static_cast<uint16_t*>(network_data) + 20, 4);
+  memcpy(&rconn.rci.rkey, static_cast<int*>(network_data) + 20, 4);
 
   printf("\n============ (received) remote setup ===============\n");
   printf("===== ADDRESS : %p\n", reinterpret_cast<void*>(rconn.rci.buf_addr));
